+++ webtools/update/developers/usermanager.php
@@ -60,96 +60,101 @@ if ($_POST["submit"] && $_GET["action"]=="update") {
 <h1>Updating User List...</h1>
 <?php
 
-//Process Post Data, Make Changes to User Table.
-//Begin General Updating
-for ($i=1; $i<=$_POST[maxuserid]; $i++) {
-$admin = $_POST["admin$i"];
-$editor = $_POST["editor$i"];
-$trusted = $_POST["trusted$i"];
-$disabled = $_POST["disabled$i"];
-//echo"$i - $admin - $editor - $trusted<br>\n";
+    //Process Post Data, Make Changes to User Table.
+    //Begin General Updating
+    for ($i=1; $i<=$_POST[maxuserid]; $i++) {
+      $admin = $_POST["admin$i"];
+      $editor = $_POST["editor$i"];
+      $trusted = $_POST["trusted$i"];
+      $disabled = $_POST["disabled$i"];
+      $selected = $_POST["selected$i"];
+      //echo "$i - $admin - $editor - $trusted - $selected<br>\n";
 
-if ($admin=="TRUE") { $mode="A"; 
-} else if ($editor=="TRUE") { $mode="E"; 
-} else if ($disabled=="TRUE") {$mode="D";
-} else { $mode="U"; }
+      if ($admin=="TRUE") { $mode="A"; 
+      } else if ($editor=="TRUE") { $mode="E"; 
+      } else if ($disabled=="TRUE") {$mode="D";
+      } else { $mode="U"; }
 
-if ($trusted !=="TRUE") {$trusted="FALSE"; }
+      if ($trusted !=="TRUE") {$trusted="FALSE"; }
 
-$sql = "UPDATE `t_userprofiles` SET `UserMode`= '$mode', `UserTrusted`= '$trusted' WHERE `UserID`='$i'";
- $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-
-}
-unset($i);
-
-echo"Your changes to the User List have been succesfully completed<BR>\n";
-
-//Do Special Disable, Delete, Enable Account Operations
-if ($_POST["selected"] AND $_POST["submit"] !=="Update") {
-$selecteduser = $_POST["selected"];
-
-if ($_POST["submit"]=="Disable Selected") {
-$sql = "UPDATE `t_userprofiles` SET `UserMode`= 'D' WHERE `UserID`='$selecteduser'";
-    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-    if ($sql_result) {
-    echo"User Account for User Number $selecteduser Disabled<br>\n";
-    }
-
-  //Disabling an author, check their extension list and disable any item they're the solo author of.
-    $sql = "SELECT TM.ID, TM.Name from `t_main` TM INNER JOIN `t_authorxref` TAX ON TM.ID=TAX.ID WHERE TAX.UserID = '$selecteduser'";
-      $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-      while ($row = mysql_fetch_array($sql_result)) {
-      $id = $row["ID"];
-      $name = $row["Name"];
-        $sql2 = "SELECT `ID` from `t_authorxref` WHERE `ID` = '$id'";
-          $sql_result2 = mysql_query($sql2, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-          if (mysql_num_rows($sql_result2)<="1") {
-           $sql3 = "UPDATE `t_version` SET `approved`='DISABLED' WHERE `ID`='$id' and `approved` !='NO' ";
-             $sql_result3 = mysql_query($sql3, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-             if ($sql_result3) {
-             echo"$name disabled from public viewing...<br>\n";
-             }
-          }
+      if (checkFormKey()) {
+        $sql = "UPDATE `t_userprofiles` SET `UserMode`= '$mode', `UserTrusted`= '$trusted' WHERE `UserID`='$i'";
+        $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
       }
 
-} else if ($_POST["submit"]=="Delete Selected") {
-$sql = "DELETE FROM `t_userprofiles` WHERE `UserID`='$selecteduser' LIMIT 1";
-    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-    if ($sql_result) {
-    echo"User Account for User Number $selecteduser Deleted<br>\n";
-    }
+      //Do Special Disable, Delete, Enable Account Operations
+      if ($_POST["selected$i"] AND $_POST["submit"] !=="Update") {
+	$selecteduser = $_POST["selected$i"];
 
-} else if ($_POST["submit"]=="Enable Selected") {
-$sql = "UPDATE `t_userprofiles` SET `UserMode`= 'U' WHERE `UserID`='$selecteduser'";
-  $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-  if ($sql_result) {
-  echo"User Account for User Number $selecteduser Enabled, User Mode set to User<br>\n";
-  }
+	if ($_POST["submit"]=="Disable Selected") {
+          if (checkFormKey()) {
+	    $sql = "UPDATE `t_userprofiles` SET `UserMode`= 'D' WHERE `UserID`='$selecteduser'";
+	    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	    if ($sql_result) {
+	      echo"User Account for User Number $selecteduser Disabled<br>\n";
+	    }
 
-  //Disabling an author, check their extension list and disable any item they're the solo author of.
-    $sql = "SELECT TM.ID, TM.Name from `t_main` TM INNER JOIN `t_authorxref` TAX ON TM.ID=TAX.ID WHERE TAX.UserID = '$selecteduser'";
-      $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-      while ($row = mysql_fetch_array($sql_result)) {
-      $id = $row["ID"];
-      $name = $row["Name"];
-        $sql2 = "SELECT `ID` from `t_authorxref` WHERE `ID` = '$id'";
-          $sql_result2 = mysql_query($sql2, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-          if (mysql_num_rows($sql_result2)<="1") {
-           $sql3 = "UPDATE `t_version` SET `approved`='?' WHERE `ID`='$id' and `approved` !='NO'";
-             $sql_result3 = mysql_query($sql3, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-             if ($sql_result3) {
-             echo"$name restored to public view pending approval...<br>\n";
-             }
+	    //Disabling an author, check their extension list and disable any item they're the solo author of.
+	    $sql = "SELECT TM.ID, TM.Name from `t_main` TM INNER JOIN `t_authorxref` TAX ON TM.ID=TAX.ID WHERE TAX.UserID = '$selecteduser'";
+	    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	    while ($row = mysql_fetch_array($sql_result)) {
+	      $id = $row["ID"];
+	      $name = $row["Name"];
+	      $sql2 = "SELECT `ID` from `t_authorxref` WHERE `ID` = '$id'";
+	      $sql_result2 = mysql_query($sql2, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	      if (mysql_num_rows($sql_result2)<="1") {
+	        $sql3 = "UPDATE `t_version` SET `approved`='DISABLED' WHERE `ID`='$id' and `approved` !='NO' ";
+	        $sql_result3 = mysql_query($sql3, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	        if ($sql_result3) {
+	  	  echo"$name disabled from public viewing...<br>\n";
+	        }
+	      }
+            }
+	  }
+
+	} else if ($_POST["submit"]=="Delete Selected") {
+          if (checkFormKey()) {
+	    $sql = "DELETE FROM `t_userprofiles` WHERE `UserID`='$selecteduser' LIMIT 1";
+	    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	    if ($sql_result) {
+	      echo"User Account for User Number $selecteduser Deleted<br>\n";
+	    }
           }
+
+	} else if ($_POST["submit"]=="Enable Selected") {
+          if (checkFormKey()) {
+	    $sql = "UPDATE `t_userprofiles` SET `UserMode`= 'U' WHERE `UserID`='$selecteduser'";
+	    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	    if ($sql_result) {
+	      echo"User Account for User Number $selecteduser Enabled, User Mode set to User<br>\n";
+	    }
+	    //Disabling an author, check their extension list and disable any item they're the solo author of.
+	    $sql = "SELECT TM.ID, TM.Name from `t_main` TM INNER JOIN `t_authorxref` TAX ON TM.ID=TAX.ID WHERE TAX.UserID = '$selecteduser'";
+	    $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	    while ($row = mysql_fetch_array($sql_result)) {
+	      $id = $row["ID"];
+	      $name = $row["Name"];
+	      $sql2 = "SELECT `ID` from `t_authorxref` WHERE `ID` = '$id'";
+	      $sql_result2 = mysql_query($sql2, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	      if (mysql_num_rows($sql_result2)<="1") {
+	        $sql3 = "UPDATE `t_version` SET `approved`='?' WHERE `ID`='$id' and `approved` !='NO'";
+	        $sql_result3 = mysql_query($sql3, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
+	        if ($sql_result3) {
+	  	  echo"$name restored to public view pending approval...<br>\n";
+	        }
+	      }
+	    }
+          }
+	}
       }
 
 
+    }
+  unset($i);
 
-}
+  echo"Your changes to the User List have been succesfully completed<BR>\n";
 
-}
-
-}
+ }
 ?>
 <h1>Manage User list</h1>
 <TABLE BORDER=0 CELLPADDING=1 CELLSPACING=1 ALIGN=CENTER STYLE="border: 0px; width: 100%" class="listing">
@@ -163,6 +168,7 @@ $sql = "UPDATE `t_userprofiles` SET `UserMode`= 'U' WHERE `UserID`='$selecteduse
 <TH>T</TH>
 </TR>
 <FORM NAME="updateusers" METHOD="POST" ACTION="?function=&action=update">
+<?writeFormKey();?>
 <?php
  $sql = "SELECT * FROM `t_userprofiles` ORDER BY `UserMode`, `UserName` ASC";
  $sql_result = mysql_query($sql, $connection) or trigger_error("MySQL Error ".mysql_errno().": ".mysql_error()."", E_USER_NOTICE);
@@ -185,7 +191,7 @@ if ($usermode=="A") {$a="TRUE"; $e="TRUE";
     echo"<TD CLASS=\"tablehighlight\" ALIGN=CENTER><B>$i</B></TD>\n";
     echo"<TD CLASS=\"tablehighlight\"><B>&nbsp;&nbsp;<A HREF=\"?function=edituser&userid=$userid\">$username</A></B></TD>\n";
     echo"<TD CLASS=\"tablehighlight\"><B>&nbsp;&nbsp;<A HREF=\"mailto:$useremail\">$useremail</A></B></TD>\n";
-    echo"<TD CLASS=\"tablehighlight\"><INPUT NAME=\"selected\" TYPE=\"RADIO\" VALUE=\"$userid\" TITLE=\"Selected User\""; if (($a=="TRUE" or $e=="TRUE") AND $_SESSION["level"]=="editor") {echo" DISABLED=\"DISABLED\"";} echo"></TD>";
+    echo"<TD CLASS=\"tablehighlight\"><INPUT NAME=\"selected$userid\" TYPE=\"CHECKBOX\" VALUE=\"$userid\" TITLE=\"Selected User\""; if (($a=="TRUE" or $e=="TRUE") AND $_SESSION["level"]=="editor") {echo" DISABLED=\"DISABLED\"";} echo"></TD>";
     echo"<TD CLASS=\"tablehighlight\"><INPUT NAME=\"editor$userid\" TYPE=\"CHECKBOX\" VALUE=\"TRUE\" "; if ($e=="TRUE") {echo"CHECKED=\"CHECKED\""; } if (($a=="TRUE" or $d=="TRUE") or $_SESSION["level"]=="editor") {echo" DISABLED=\"DISABLED\"";} echo" TITLE=\"Editor\"></TD>";
     echo"<TD CLASS=\"tablehighlight\"><INPUT NAME=\"admin$userid\" TYPE=\"CHECKBOX\" VALUE=\"TRUE\" "; if ($a=="TRUE") {echo"CHECKED=\"CHECKED\""; } if ($d=="TRUE" or $_SESSION["level"]=="editor") {echo" DISABLED=\"DISABLED\"";} echo" TITLE=\"Administrator\"></TD>";
     echo"<TD CLASS=\"tablehighlight\"><INPUT NAME=\"trusted$userid\" TYPE=\"CHECKBOX\" VALUE=\"TRUE\" "; if ($t=="TRUE") {echo"CHECKED=\"CHECKED\""; } if ($d=="TRUE" or (($a=="TRUE" or $e=="TRUE") AND $_SESSION["level"]=="editor" )) {echo" DISABLED=\"DISABLED\"";}echo" TITLE=\"Trusted User\"></TD>";
@@ -204,7 +210,7 @@ echo"<INPUT NAME=\"maxuserid\" TYPE=\"HIDDEN\" VALUE=\"$maxuserid\">";
 <INPUT NAME="submit" TYPE="SUBMIT" VALUE="Delete Selected" ONCLICK="return confirm('Deleting this account will permanently remove all their extensions and account information from Mozilla Update. This cannot be undone. Do you want to procede and delete this account?');">
 <INPUT NAME="submit" TYPE="SUBMIT" VALUE="Update">
 </TD>
-<TD><INPUT NAME="selected" TYPE="RADIO" VALUE="" CHECKED TITLE="Selected User"></TD>
+<TD>
 </TR>
 </FORM>
 </TABLE>
@@ -212,6 +218,7 @@ echo"<INPUT NAME=\"maxuserid\" TYPE=\"HIDDEN\" VALUE=\"$maxuserid\">";
 <div style="width: 580px; border: 0px dotted #AAA; margin-top: 1px; margin-left: 50px; margin-bottom: 5px; font-size: 10pt; font-weight: bold">
 
 <form name="adduser" method="post" action="?function=adduser">
+<?writeFormKey();?>
   E-Mail: <input name="email" type="text" size="30" maxlength="150" value="">
 <input name="submit" type="submit" value="Add User"></SPAN>
 </form>
@@ -248,17 +255,20 @@ $useremail = $_POST["useremail"];
 $userwebsite = $_POST["userwebsite"];
 $useremailhide = $_POST["useremailhide"];
 
-  $sql = "UPDATE `t_userprofiles` SET `UserName`= '$username', `UserEmail`='$useremail', `UserWebsite`='$userwebsite', `UserMode`='$mode', `UserTrusted`='$trusted', `UserEmailHide`='$useremailhide' WHERE `UserID`='$userid'";
+  if (checkFormKey()) {
+    $sql = "UPDATE `t_userprofiles` SET `UserName`= '$username', `UserEmail`='$useremail', `UserWebsite`='$userwebsite', `UserMode`='$mode', `UserTrusted`='$trusted', `UserEmailHide`='$useremailhide' WHERE `UserID`='$userid'";
     $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
     if ($sql_result) {
     echo"<h1>Updating User Profile...</h1>\n";
     echo"The User Profile for $username, has been successfully updated...<br>\n";
     }
+  }
 } else if ($_POST["submit"] == "Delete User") {
 if ($_SESSION["level"] !=="admin" && $_SESSION["uid"] !== $_POST["userid"]) {$_POST["userid"]=$_SESSION["uid"];}
 $userid = $_POST["userid"];
 $username = $_POST["username"];
-  $sql = "DELETE FROM `t_userprofiles` WHERE `UserID`='$userid'";
+  if (checkFormKey()) {
+    $sql = "DELETE FROM `t_userprofiles` WHERE `UserID`='$userid'";
     $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
     if ($sql_result) {
     echo"<h1>Deleting User... Please wait...</h1>\n";
@@ -267,6 +277,7 @@ $username = $_POST["username"];
     echo"</body>\n</html>\n";
     exit;
     }
+  }
 }
 
 if (!$userid) {$userid=$_POST["userid"];}
@@ -290,6 +301,7 @@ if (!$userid) {$userid=$_POST["userid"];}
 <TABLE BORDER=0 CELLPADDING=2 CELLSPACING=2 ALIGN=CENTER STYLE="border: 0px; width: 95%">
 <TR><TD COLSPAN=2>Last login: <?php echo"$userlastlogin"; ?></TD></TR>
 <FORM NAME="edituser" METHOD="POST" ACTION="?function=edituser">
+<?writeFormKey();?>
 <?php
     echo"<INPUT NAME=\"userid\" TYPE=\"HIDDEN\" VALUE=\"$userid\">\n";
     echo"<TR><TD STYLE=\"width: 130px\"><B>Name:</B></TD><TD><INPUT NAME=\"username\" TYPE=\"TEXT\" VALUE=\"$username\" SIZE=30 MAXLENGTH=100></TD></TR>\n";
@@ -375,13 +387,15 @@ $useremail = $_POST[useremail];
 $userwebsite = $_POST[userwebsite];
 $userpass = $_POST[userpass];
 $useremailhide = $_POST[useremailhide];
-   	$sql = "INSERT INTO `t_userprofiles` (`UserName`, `UserEmail`, `UserWebsite`, `UserPass`, `UserMode`, `UserTrusted`, `UserEmailHide`) VALUES ('$username', '$useremail', '$userwebsite', '$userpass', '$mode', '$trusted', '$useremailhide');";
+  if (checkFormKey()) {
+    $sql = "INSERT INTO `t_userprofiles` (`UserName`, `UserEmail`, `UserWebsite`, `UserPass`, `UserMode`, `UserTrusted`, `UserEmailHide`) VALUES ('$username', '$useremail', '$userwebsite', '$userpass', '$mode', '$trusted', '$useremailhide');";
     $sql_result = mysql_query($sql) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
     if ($sql_result) {
       include"mail_newaccount.php"; 
       echo"The user $username has been added successfully...<br>\n";
       echo"An E-Mail has been sent to the e-mail address specified with the login info they need to log in to their new account.<br>\n";
     }
+  }
 }
 
 }
@@ -390,6 +404,7 @@ $useremailhide = $_POST[useremailhide];
 <h1>Add New User</h1>
 <TABLE BORDER=0 CELLPADDING=2 CELLSPACING=2 ALIGN=CENTER STYLE="border: 0px; width: 95%">
 <FORM NAME="adduser" METHOD="POST" ACTION="?function=adduser">
+<?writeFormKey();?>
     <TR><TD><B>E-Mail:</B></TD><TD><INPUT NAME="useremail" TYPE="TEXT" VALUE="<?php echo"$_POST[email]"; ?>" SIZE=30 MAXLENGTH=100></TD></TR>
     <TR><TD ALIGN=RIGHT><B>Show E-Mail:<B></TD><TD>Hidden: <INPUT NAME="useremailhide" TYPE="RADIO" VALUE="1" CHECKED> Visible: <INPUT NAME="useremailhide" TYPE="RADIO" VALUE="0"></TD></TR>
     <TR><TD STYLE="width: 150px"><B>Name:</B></TD><TD><INPUT NAME="username" TYPE="TEXT" VALUE="" SIZE=30 MAXLENGTH=100></TD></TR>
@@ -424,12 +439,15 @@ if ($_POST["submit"]=="Change Password") {
      $password_plain = $newpassword;
      $userpass = md5($newpassword);
 
-     $sql = "UPDATE `t_userprofiles` SET `UserPass`='$userpass' WHERE `UserID`='$userid'";
+     $sql_result = false;
+     if (checkFormKey()) {
+       $sql = "UPDATE `t_userprofiles` SET `UserPass`='$userpass' WHERE `UserID`='$userid'";
        $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
-       if ($sql_result) {
+     }
+     if ($sql_result) {
        include"mail_newpassword.php";
        echo"The password has been successfully changed, an e-mail has been sent confirming this action.<br>\n";
-       }
+     }
     } else {
      echo"The two passwords did not match, please go back and try again.<BR>\n";
     }
@@ -452,13 +470,14 @@ if ($_POST["submit"]=="Change Password") {
        $row = mysql_fetch_array($sql_result);
          $email = $row["UserEmail"];
 
-     $sql = "UPDATE `t_userprofiles` SET `UserPass`='$userpass' WHERE `UserID`='$userid'";
+     if (checkFormKey()) {
+       $sql = "UPDATE `t_userprofiles` SET `UserPass`='$userpass' WHERE `UserID`='$userid'";
        $sql_result = mysql_query($sql, $connection) or trigger_error("<FONT COLOR=\"#FF0000\"><B>MySQL Error ".mysql_errno().": ".mysql_error()."</B></FONT>", E_USER_NOTICE);
        if ($sql_result) {
        include"mail_newpassword.php";
        echo"The password has been successfully reset. The user has been sent an e-mail notifying them of their new password.<br>\n";
        }
-   
+     }
 }
 
 if (!$userid) { $userid = $_POST["userid"]; }
@@ -471,6 +490,7 @@ if (!$userid) { $userid = $_POST["userid"]; }
 <h1>Change password for <?php echo"$username"; ?></h1>
 <TABLE BORDER=0 CELLPADDING=2 CELLSPACING=2 ALIGN=CENTER STYLE="border: 0px; width: 95%">
 <FORM NAME="adduser" METHOD="POST" ACTION="?function=changepassword&userid=<?php echo"$userid"; ?>">
+<?writeFormKey();?>
     <INPUT NAME="userid" TYPE="HIDDEN" VALUE="<?php echo"$userid"; ?>">
 <?php if (($_SESSION["level"] =="admin" or $_SESSION["level"]=="editor") and $userid != $_SESSION["uid"]) { ?>
     <TR><TD COLSPAN="2" ALIGN="CENTER"><INPUT NAME="submit" TYPE="SUBMIT" VALUE="Generate New Password"></TD></TR>
